+++ holiday-manager-core/src/main/java/holiday/manager/port/adapter/persistence/repository/holiday/EventStoreHolidayApplicationRepository.java
@@ -3,9 +3,9 @@ package holiday.manager.port.adapter.persistence.repository.holiday;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import holiday.manager.domain.model.holiday.HolidayApplication;
-import holiday.manager.domain.model.holiday.HolidayApplicationId;
-import holiday.manager.domain.model.holiday.HolidayApplicationRepository;
+import holiday.manager.domain.model.holiday.application.HolidayApplication;
+import holiday.manager.domain.model.holiday.application.HolidayApplicationId;
+import holiday.manager.domain.model.holiday.application.HolidayApplicationRepository;
 import holiday.manager.event.sourcing.EventStore;
 import holiday.manager.event.sourcing.EventStream;
 import holiday.manager.event.sourcing.EventStreamId;
