+++ holiday-manager-core/src/test/java/holiday/manager/application/HolidayApplicationServiceTest.java
@@ -17,15 +17,15 @@ import holiday.manager.application.query.holiday.HolidayApplicationQueryService;
 import holiday.manager.application.query.holiday.dto.HolidayApplicationDto;
 import holiday.manager.application.service.holiday.HolidayApplicationService;
 import holiday.manager.domain.model.AuthenticationException;
-import holiday.manager.domain.model.holiday.HolidayApplication;
-import holiday.manager.domain.model.holiday.HolidayApplicationRepository;
-import holiday.manager.domain.model.holiday.HolidayApplicationStatus;
-import holiday.manager.domain.model.holiday.HolidayType;
-import holiday.manager.domain.model.holiday.KindOfHoliday;
-import holiday.manager.domain.model.holiday.event.HolidayApplicationApplied;
-import holiday.manager.domain.model.holiday.event.HolidayApplicationApproved;
-import holiday.manager.domain.model.holiday.event.HolidayApplicationCanceled;
-import holiday.manager.domain.model.holiday.event.HolidayApplicationRejected;
+import holiday.manager.domain.model.holiday.application.HolidayApplication;
+import holiday.manager.domain.model.holiday.application.HolidayApplicationRepository;
+import holiday.manager.domain.model.holiday.application.HolidayApplicationStatus;
+import holiday.manager.domain.model.holiday.application.HolidayType;
+import holiday.manager.domain.model.holiday.application.KindOfHoliday;
+import holiday.manager.domain.model.holiday.application.event.HolidayApplicationApplied;
+import holiday.manager.domain.model.holiday.application.event.HolidayApplicationApproved;
+import holiday.manager.domain.model.holiday.application.event.HolidayApplicationCanceled;
+import holiday.manager.domain.model.holiday.application.event.HolidayApplicationRejected;
 import holiday.manager.domain.model.user.User;
 import holiday.manager.domain.model.user.UserId;
 import holiday.manager.event.sourcing.EventStore;
